Create SnakeGame
Direction Input: After running the Snake Game program, it will display the game board and prompt you to enter a direction using the keys W, A, S, and D. These keys correspond to the following directions:

W: Move the snake up
A: Move the snake left
S: Move the snake down
D: Move the snake right
Playing the Game:

Enter one of the direction keys (W, A, S, or D) followed by pressing the Enter key to move the snake in the desired direction.
Continue entering directions to control the snake's movement and try to eat the food (*) to grow longer.
The game will end if the snake goes out of bounds or collides with itself.
Game Over Message: When the game ends, either due to a collision or going out of bounds, the program will display a "Game Over" message with the reason for the game ending (e.g., "Game Over! Out of bounds." or "Game Over! Self collision.").

Restarting the Game: To play again, you can rerun the Snake Game program, and it will start a new game session.

Signed-off-by: Yashitha29 <151449950+Yashitha29@users.noreply.github.com>